requirements.txt
absl-py==1.4.0
addict==2.4.0
aiohttp==3.8.5
aiosignal==1.3.1
aist-plusplus-api==1.1.0
aliyun-python-sdk-core==2.14.0
aliyun-python-sdk-kms==2.16.2
aniposelib==0.4.3
async-timeout==4.0.3
attrs==23.1.0
audioread==3.0.0
beautifulsoup4==4.12.3
bs4==0.0.2
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
chumpy==0.70
click==8.1.7
clip @ git+https://github.com/openai/CLIP.git@a1d071733d7111c9c014f024669f959182114e33
colorama==0.4.6
contourpy==1.1.0
crcmod==1.7
cryptography==41.0.4
cycler==0.11.0
Cython==3.0.6
datasets==2.14.4
decorator==4.4.2
decord==0.6.0
Deprecated==1.2.14
dill==0.3.7
easydict==1.10
einops==0.6.1
essentia==2.1b6.dev1034
et-xmlfile==1.1.0
fastdtw==0.3.4
ffmpeg-python==0.2.0
filelock==3.12.3
flatbuffers==23.5.26
fonttools==4.42.1
frozenlist==1.4.0
fsspec==2023.9.0
ftfy==6.1.1
future==0.18.3
fvcore==0.1.5.post20221221
google-auth==2.22.0
google-auth-oauthlib==1.0.0
grpcio==1.57.0
huggingface-hub==0.16.4
idna==3.4
imageio==2.31.2
imageio-ffmpeg==0.4.8
importlib-metadata==6.8.0
importlib-resources==6.0.1
iopath==0.1.10
jmespath==0.10.0
joblib==1.3.2
keras==2.13.1
kiwisolver==1.4.5
lazy_loader==0.3
librosa==0.10.1
llvmlite==0.40.1
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.2
mdurl==0.1.2
mediapipe==0.10.3
mmcv==2.0.0
mmengine==0.9.0
model-index==0.1.11
moviepy==1.0.3
msgpack==1.0.5
multidict==6.0.4
multiprocess==0.70.15
numba==0.57.1
numpy==1.23.0
oauthlib==3.2.2
onnx==1.15.0
opencv-contrib-python==4.8.0.76
opencv-python==4.8.0.76
opendatalab==0.0.10
openmim==0.3.9
openpose-pytorch==0.1.0
openpyxl==3.1.2
openxlab==0.0.27
ordered-set==4.1.0
oss2==2.17.0
packaging==23.1
pandas==2.0.3
panns-inference==0.1.1
Pillow==10.0.0
platformdirs==3.10.0
pooch==1.7.0
portalocker==2.8.2
proglog==0.1.10
protobuf==3.20.3
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.19.0
Pygments==2.16.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-fcl==0.7.0.5
pytorch-fast-transformers==0.4.0
pytorch3d==0.7.5
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.8.8
requests==2.28.2
requests-oauthlib==1.3.1
rich==13.4.2
rsa==4.9
safetensors==0.3.3
scikit-learn==1.3.0
scipy==1.10.1
seaborn==0.13.0
six==1.16.0
smplpytorch==0.0.8
smplx==0.1.28
sounddevice==0.4.6
soundfile==0.12.1
soupsieve==2.5
soxr==0.3.6
tabulate==0.9.0
tensorboard==2.14.0
tensorboard-data-server==0.7.1
tensorboardX==2.6.2.2
termcolor==2.3.0
threadpoolctl==3.2.0
timm==0.9.7
tokenizers==0.13.3
toml==0.10.2
tomli==2.0.1
torch==1.13.1+cu116
torchaudio==0.13.1+cu116
torchlibrosa==0.1.0
torchvision==0.14.1+cu116
tqdm==4.65.2
transformers==4.33.0
trimesh==3.23.5
typing_extensions==4.7.1
tzdata==2023.3
ujson==5.8.0
urllib3==1.26.16
vedo==2024.5.1
vtk==9.2.6
wcwidth==0.2.6
Werkzeug==2.3.7
wrapt==1.15.0
xrmogen==0.4.0
xrprimer==0.7.1
xxhash==3.3.0
yacs==0.1.8
yapf==0.40.2
yarl==1.9.2
you-get==0.4.1650
zipp==3.16.2
